pyproject.toml
[tool.black]
line-length = 240

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "oryx"
version = "1.0.0"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
]

dependencies = [
    "torch", "torchvision",
    "transformers==4.39.2", "tokenizers==0.15.2", "sentencepiece==0.1.99", "shortuuid",
    "accelerate==0.27.2", "peft==0.4.0", "bitsandbytes==0.41.0",
    "pydantic<2,>=1", "markdown2[all]", "numpy", "scikit-learn==1.2.2",
    "gradio==3.35.2", "gradio_client==0.2.9",
    "requests", "httpx==0.24.0", "uvicorn", "fastapi",
    "einops==0.6.1", "einops-exts==0.0.4", "timm==0.9.16", "decord","ninja","deepspeed==0.12.2","protobuf", 
    "pythainlp", "datasets", "pillow"
]

[tool.setuptools.packages.find]
exclude = [
    "assets*",
    "benchmark*",
    "docs",
    "dist*",
    "playground*",
    "scripts*",
    "tests*",
    "checkpoints*",
    "project_checkpoints*",
    "work_dirs*",
    "data*",
]

[tool.wheel]
exclude = [
    "assets*",
    "benchmark*",
    "docs",
    "dist*",
    "playground*",
    "scripts*",
    "tests*",
    "checkpoints*",
    "project_checkpoints*",
    "work_dirs*",
    "data*",
]